2. adk/mal_season_agent/agent.py
from google.adk.agents import LlmAgent
from . import mal_scraper_tool

root_agent = LlmAgent(
    name="MALSeasonCheckerAgent",
    description="Agen AI untuk menampilkan daftar anime yang sedang tayang berdasarkan musim dari MyAnimeList.net.",
    model="gemini-2.0-flash", 
    instruction=(
        "Anda adalah MALSeasonChecker, sebuah agen AI yang bertugas memberikan informasi tentang anime apa saja yang sedang tayang di musim ini (atau musim tertentu jika diminta) berdasarkan data dari MyAnimeList.net.\n"
        "Gunakan alat `Workspace_seasonal_anime_from_mal` untuk mendapatkan informasi ini.\n"
        "Jika pengguna hanya bertanya 'anime musim ini' atau serupa, gunakan musim dan tahun saat ini secara otomatis.\n"
        "Jika pengguna menyebutkan musim dan tahun spesifik (misalnya 'anime musim gugur 2024'), gunakan informasi tersebut untuk alat.\n"
        "Sajikan hasilnya dalam format yang mudah dibaca.\n"
        "Alat mungkin mengembalikan banyak anime, jadi sampaikan ringkasan atau beberapa yang teratas jika responsnya terlalu panjang, atau tanyakan apakah pengguna ingin melihat lebih banyak."
        "Konfirmasi musim dan tahun yang Anda gunakan untuk pencarian."
    ),
    tools=[
        mal_scraper_tool.fetch_seasonal_anime_from_mal,
    ]
)
