2. adk/pyproject.toml
[project]
name = "2-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-adk>=0.5.0",
    "requests>=2.32.3",
]
